Move logdata header files
