+++ src/logdata/CMakeLists.txt
@@ -1,16 +1,16 @@
 add_library(logdata STATIC
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/abstractlogdata.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/blockpool.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/compressedlinestorage.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/encodingdetector.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/linepositionarray.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/loadingstatus.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/logdata.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/logdataworkerthread.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/logfiltereddata.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/logfiltereddataworkerthread.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/marks.h
-         ${CMAKE_CURRENT_SOURCE_DIR}/include/data/linetypes.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/abstractlogdata.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/blockpool.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/compressedlinestorage.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/encodingdetector.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/linepositionarray.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/loadingstatus.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/logdata.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/logdataworkerthread.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/logfiltereddata.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/logfiltereddataworkerthread.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/marks.h
+         ${CMAKE_CURRENT_SOURCE_DIR}/include/linetypes.h
 
         ${CMAKE_CURRENT_SOURCE_DIR}/src/abstractlogdata.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/src/blockpool.cpp
@@ -25,7 +25,6 @@ add_library(logdata STATIC
 )
 
 target_include_directories(logdata PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/include")
-target_include_directories(logdata PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include/data")
 
 target_link_libraries(logdata PUBLIC Qt5::Core Qt5::Concurrent tbb_static moodycamel named_type absl::optional utils filewatch uchardet)
 set_property(TARGET logdata PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
