+++ src/logdata/include/abstractlogdata.h
