+++ src/logdata/include/compressedlinestorage.h
