+++ src/logdata/include/encodingdetector.h
