+++ src/logdata/include/linepositionarray.h
