+++ src/logdata/include/loadingstatus.h
