+++ src/logdata/include/logdataworkerthread.h
