+++ src/logdata/include/logfiltereddata.h
