+++ src/logdata/include/logfiltereddataworkerthread.h
