+++ src/ui/src/quickfind.cpp
@@ -47,7 +47,7 @@
 #include "log.h"
 #include "quickfindpattern.h"
 #include "selection.h"
-#include "data/abstractlogdata.h"
+#include "abstractlogdata.h"
 
 #include "quickfind.h"
