+++ src/ui/src/session.cpp
@@ -29,8 +29,8 @@
 #include "persistentinfo.h"
 #include "savedsearches.h"
 #include "sessioninfo.h"
-#include "data/logdata.h"
-#include "data/logfiltereddata.h"
+#include "logdata.h"
+#include "logfiltereddata.h"
 
 Session::Session()
 {
