+++ tests/logfiltereddataPerfTest.cpp
@@ -4,8 +4,8 @@
 #include "log.h"
 #include "test_utils.h"
 
-#include "data/logdata.h"
-#include "data/logfiltereddata.h"
+#include "logdata.h"
+#include "logfiltereddata.h"
 
 #include "gmock/gmock.h"
