+++ tests/logfiltereddataTest.cpp
@@ -26,8 +26,8 @@
 #include "persistentinfo.h"
 #include "test_utils.h"
 
-#include "data/logdata.h"
-#include "data/logfiltereddata.h"
+#include "logdata.h"
+#include "logfiltereddata.h"
 
 #include <catch.hpp>
